requirements.txt
wheel
pillow
beautifulsoup4
urllib3
numpy
pyoxipng